Extract ConnectionDialog field-input helpers into a mixin
Move the autocomplete/picker helpers (jump-host and tag picker popovers,
comma-separated inline completion, and the Wake-on-LAN MAC auto-detect
button) out of the ConnectionDialog god-object into a new
ConnectionDialogFieldHelpersMixin in connection_dialog_field_helpers.py.

The move is verbatim: ConnectionDialog inherits the mixin so self stays
identical and no logic changes. threading and the wol import were only used
by this cluster, so they move with it and are dropped from connection_dialog.py.

Add focused tests for the pure candidate/text helpers (tag/jump-host
candidate lists and set-text-without-completion) plus an extraction-contract
guard, none of which had coverage before.
